registro_frequencia_aluno usando contexto de consulta
