+++ src/SME.SGP.Dominio.Interfaces/Repositorios/IRepositorioConsolidacaoFrequenciaTurmaConsulta.cs
@@ -1,0 +1,14 @@
+﻿using SME.SGP.Infra;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace SME.SGP.Dominio.Interfaces
+{
+    public interface IRepositorioConsolidacaoFrequenciaTurmaConsulta
+    {
+        Task<IEnumerable<GraficoAusenciasComJustificativaDto>> ObterAusenciasComJustificativaASync(int anoLetivo, long dreId, long ueId, Modalidade? modalidade, int semestre);
+        Task<IEnumerable<FrequenciaGlobalPorAnoDto>> ObterFrequenciaGlobalPorAnoAsync(int anoLetivo, long dreId, long ueId, Modalidade? modalidade, int semestre);
+        Task<IEnumerable<FrequenciaGlobalPorDreDto>> ObterFrequenciaGlobalPorDreAsync(int anoLetivo, Modalidade modalidade, string ano, int? semestre);
+        Task<bool> ExisteConsolidacaoFrequenciaTurmaPorAno(int ano);        
+    }
+}
